Content & Responsiveness
The text content has been added to both the Security Overlay and the Sound Overlay. Additions to other overlays will come later, as the finished texts are not yet available. The curtain string has been tweaked to better suit different screens, and a few elements are a bit more responsive. There is also an unfinished attempt to implement the WiFi-button.