Cleanup & Polish
The tile icons have been adjusted slightly, to be more consistent in terms of size and margins. The blurred background now always covers the entire screen, without adding extra scroll to the page. Several pictures are now actually transparent, and the WiFi-overlay has been tuned to be built in the same manner as the other overlays.